Reference

Public API

RelationalEvents.RelationalEvent — Type

RelationalEvent(sender, receiver, time)

Type to represent a basic unmarked relational event containing the sender, the receiver, and the timestamp of the event.

Sender and receiver are required to be Node{T}, where T can be any type. Nodes carry a contiguous integer id alongside the node data.

Examples

julia> sender = RelationalEvents.Node(1, "a");

julia> receiver = RelationalEvents.Node(2, "b");

julia> time = 1.0;

julia> RelationalEvent(sender, receiver, time)
RelationalEvent{String, Float64}
 sender: a
 receiver: b
 time: 1.0

RelationalEvents.MarkedRelationalEvent — Type

MarkedRelationalEvent(sender, receiver, time, mark)

Type to represent a marked relational event containing the sender, the receiver, the timestamp, and the mark of the event.

Sender and receiver are required to be Node{T}, where T can be any type. Nodes carry a contiguous integer id alongside the node data.

Examples

julia> sender = RelationalEvents.Node(1, "a");

julia> receiver = RelationalEvents.Node(2, "b");

julia> time = 1.0;

julia> mark = "x";

julia> MarkedRelationalEvent(sender, receiver, time, mark) RelationalEvent{String, Float64, String} sender: a receiver: b time: 1.0 mark: x ```

RelationalEvents.EventHistory — Type

EventHistory(events, nodes, spells)
EventHistory(events)

Type representing a relational event history. This holds a sorted list of events and the nodes that appear throughout the observation period, as well as their activity spells.

RelationalEvents.Spec — Type

Spec{T,S}

Struct containing the specification for the computation of event statistics.

Fields

N_events::Int # number of events to sample

N_cases::Int # number of control cases to sample

t0::S # Startpoint of the event history

thalf::T # halflife time for exponential decay

tol::Float64 = 0.01 # tolerance at which to set weights to zero

startmult::Float64 = 2.0 # halflife multiplier after which to start sampling

RelationalEvents.statistics — Function

statistics(h, spec; statfuns...)

Compute the statistics provided as keyword arguments for the event history h according to the specification in spec. Statistics are only computed for a sample of the specified sizes of events and control cases. Returns an EventStats object.

Arguments

h::EventHistory: The event history for which to compute statistics.

spec::Spec: The specification for sampling, decay, etc.

statfuns::Function...: Functions to compute event history statistics.

Examples

hist = ... # EventHistory
spec = Spec(
    50,   # number of events to sample
    10,   # number of control cases to sample
    1,    # origin time of the event history
    30,   # halflife time (e.g., days)
    0.01, # threshold at which to set dyads to zero
    2     # halflife multiplier after which to start sampling
)

res = statistics(hist, spec; inertia, reciprocity)

RelationalEvents.isactive — Function

isactive(h, i, t)

Check if the ith actor is active at time t. Note that when nodes nodes(h) are represented by a range of contiguous integers, i is equal to nodes(h)[i].

RelationalEvents.inertia — Function

inertia(e, p, h, spec)

Compute the inertia statistic for event e, using the EventProcess p and specification spec.

RelationalEvents.reciprocity — Function

reciprocity(e, p, h, spec)

Compute the reciprocity statistic for event e, using the EventProcess p and specification spec.

Internals

RelationalEvents.EventStats — Type

Struct containing statistics about a sampled event history and control cases, to be used for fitting a relational event model.

RelationalEvents.EventProcess — Type

EventProcess{W,E}

Struct holding sparse arrays with weights and weight update times, and vectors with in- and outdegrees. This is updated iterativeley as statistics are computed.

RelationalEvents.Node — Type

Wrapper struct containing node metadata and a contiguous integer index.

RelationalEvents.update_process! — Function

update_process!(p, e, spec)

Update weights and time-of-last-update of EventProcess p for the dyad and event time given by e and the specification spec.

RelationalEvents.update_weights! — Function

update_weights!(p, e, spec)

Apply exponential decay to the weight of the dyad given by e based on the halflife time specified by spec. If the weight is below given tolerance spec.tol, the weight will be set to zero for reasons of memory efficiency.

RelationalEvents.update_wutimes! — Function

update_wutimes!(p, e)

Set the time of the last weight update in EventProcess p to eventtime(e).

RelationalEvents.add_event! — Function

add_event!(p, e, spec)

Update EventProcess p according to specification spec for the occurred event e.